Refactor : move column contents
