+++ app/components/contents/ColumnList.tsx
@@ -1,13 +1,13 @@
 import { ColumnListBlockObjectResponse } from "@notionhq/client/build/src/api-endpoints";
 
 import { fetchBlocks } from "@/app/lib/databases";
-import Block from "../Block";
+import Block from "./Block";
 
 export default async function ColumnList({ block }: { block: ColumnListBlockObjectResponse }) {
     const children = block.has_children ? await fetchBlocks(block.id) : null;
 
     return (
-        <div className="article-content--column-list">
+        <div className="article-content--column-list" style={{ gridTemplateColumns: `repeat(${children?.items.length}, minmax(0, 1fr))` }}>
             {children?.items.map((item, index) =>
                 <Block key={index} block={item} blocks={children?.items} index={index} />)}
         </div>
